+++ nftnl/src/expr/immediate.rs
@@ -47,7 +47,7 @@ impl Verdict {
 }
 
 impl Expression for Verdict {
-    fn to_expr(&self) -> crate::Result<*mut sys::nftnl_expr> {
+    fn to_expr(&self) -> *mut sys::nftnl_expr {
         unsafe {
             let expr = try_alloc!(sys::nftnl_expr_alloc(
                 b"immediate\0" as *const _ as *const c_char
@@ -68,7 +68,7 @@ impl Expression for Verdict {
                 self.verdict_const() as u32,
             );
 
-            Ok(expr)
+            expr
         }
     }
 }