+++ nftnl/src/expr/lookup.rs
@@ -19,7 +19,7 @@ impl Lookup {
 }
 
 impl Expression for Lookup {
-    fn to_expr(&self) -> crate::Result<*mut sys::nftnl_expr> {
+    fn to_expr(&self) -> *mut sys::nftnl_expr {
         unsafe {
             let expr = try_alloc!(sys::nftnl_expr_alloc(
                 b"lookup\0" as *const _ as *const c_char
@@ -43,7 +43,7 @@ impl Expression for Lookup {
             //         libc::NFT_LOOKUP_F_INV as u32);
             // }
 
-            Ok(expr)
+            expr
         }
     }
 }