+++ nftnl/src/expr/mod.rs
@@ -9,7 +9,7 @@ use nftnl_sys as sys;
 pub trait Expression {
     /// Allocates and returns the low level `nftnl_expr` representation of this expression.
     /// The caller to this method is responsible for freeing the expression.
-    fn to_expr(&self) -> crate::Result<*mut sys::nftnl_expr>;
+    fn to_expr(&self) -> *mut sys::nftnl_expr;
 }
 
 mod bitwise;