+++ nftnl/src/lib.rs
@@ -41,19 +41,11 @@ extern crate log;
 pub use nftnl_sys;
 use nftnl_sys::libc::c_void;
 
-
-pub type Result<T> = std::result::Result<T, AllocationError>;
-
-/// Unable to allocate memory
-#[derive(err_derive::Error, Debug)]
-#[error(display = "Unable to allocate memory")]
-pub struct AllocationError(());
-
 macro_rules! try_alloc {
     ($e:expr) => {{
         let ptr = $e;
         if ptr.is_null() {
-            return Err($crate::AllocationError(()));
+            panic!("Unable to allocate memory in libnftnl");
         }
         ptr
     }};