+++ nftnl/src/table.rs
@@ -1,4 +1,4 @@
-use crate::{MsgType, ProtoFamily, Result};
+use crate::{MsgType, ProtoFamily};
 use nftnl_sys::{
     self as sys,
     libc::{self, c_void},
@@ -20,13 +20,13 @@ pub struct Table {
 
 impl Table {
     /// Creates a new table instance with the given name and protocol family.
-    pub fn new<T: AsRef<CStr>>(name: &T, family: ProtoFamily) -> Result<Table> {
+    pub fn new<T: AsRef<CStr>>(name: &T, family: ProtoFamily) -> Table {
         unsafe {
             let table = try_alloc!(sys::nftnl_table_alloc());
 
             sys::nftnl_table_set_str(table, sys::NFTNL_TABLE_NAME as u16, name.as_ref().as_ptr());
             sys::nftnl_table_set_u32(table, sys::NFTNL_TABLE_FAMILY as u16, family as u32);
-            Ok(Table { table, family })
+            Table { table, family }
         }
     }